+++ .dialyzer_ignore.exs
@@ -1,4 +1,5 @@
 [
-  ~r|lib/recode/cli_formatter.ex:.*:no_return|
+  ~r|lib/recode/cli_formatter.ex:.*:no_return|,
+  ~r|lib/recode/cli_formatter.ex:.*:unused_fun|
 ]