+++ test/recode/cli_formatter_test.exs
@@ -22,10 +22,11 @@ defmodule Recode.CLIFormatterTest do
 
     output =
       capture_io(fn ->
-        format({:tasks_finished, project})
+        format({:tasks_finished, project, 10_000})
       end)
 
     assert output |> strip_esc_seq() |> String.trim_leading() == """
+           Executed 0 tasks in 0.01s.
            Files: 1 (.ex: 1)
             Everything ok
 
@@ -48,7 +49,11 @@ defmodule Recode.CLIFormatterTest do
 
     output =
       capture_io(fn ->
-        format({:tasks_finished, project})
+        format({:tasks_finished, project, 10_000},
+          times: [{TaTask, 1}, {DaTask, 2}],
+          corrector_tasks_finished: 20_000,
+          checker_tasks_finished: 12_345
+        )
       end)
 
     assert output |> strip_esc_seq() |> String.trim_leading() == """
@@ -61,6 +66,7 @@ defmodule Recode.CLIFormatterTest do
            3 3   |end
            4 4   |
 
+           Executed 2 tasks in 0.01s.
            Files: 1 (.ex: 1)
            Updated 1 file
             Everything ok
@@ -89,7 +95,7 @@ defmodule Recode.CLIFormatterTest do
 
     output =
       capture_io(fn ->
-        format({:tasks_finished, project})
+        format({:tasks_finished, project, 10_000})
       end)
 
     assert output |> strip_esc_seq() |> String.trim_leading() == """
@@ -111,6 +117,7 @@ defmodule Recode.CLIFormatterTest do
            3 3   |end
            4 4   |
 
+           Executed 0 tasks in 0.01s.
            Files: 2 (.ex: 2)
            Updated 2 files
             Everything ok
@@ -135,7 +142,7 @@ defmodule Recode.CLIFormatterTest do
 
     output =
       capture_io(fn ->
-        format({:tasks_finished, project})
+        format({:tasks_finished, project, 10_000})
       end)
 
     assert output |> strip_esc_seq() |> String.trim_leading() == """
@@ -143,6 +150,7 @@ defmodule Recode.CLIFormatterTest do
            Updates: 2
            Changed by: test, test
 
+           Executed 0 tasks in 0.01s.
            Files: 1 (.ex: 1)
            Updated 1 file
             Everything ok
@@ -180,7 +188,7 @@ defmodule Recode.CLIFormatterTest do
 
     output =
       capture_io(fn ->
-        format({:tasks_finished, project})
+        format({:tasks_finished, project, 10_000})
       end)
 
     output = strip_esc_seq(output)
@@ -204,7 +212,7 @@ defmodule Recode.CLIFormatterTest do
 
     output =
       capture_io(fn ->
-        format({:tasks_finished, project})
+        format({:tasks_finished, project, 10_000})
       end)
 
     assert output |> strip_esc_seq() |> String.trim_leading() == """
@@ -213,6 +221,7 @@ defmodule Recode.CLIFormatterTest do
            Changed by: test
            Moved from: foo.ex
 
+           Executed 0 tasks in 0.01s.
            Files: 1 (.ex: 1)
            Moved 1 file
             Everything ok
@@ -232,13 +241,14 @@ defmodule Recode.CLIFormatterTest do
 
     output =
       capture_io(fn ->
-        format({:tasks_finished, project})
+        format({:tasks_finished, project, 10_000})
       end)
 
     assert output |> strip_esc_seq() |> String.trim_leading() == """
            File: foo.ex
            New file
 
+           Executed 0 tasks in 0.01s.
            Files: 1 (.ex: 1)
            Created 1 file
             Everything ok
@@ -258,13 +268,14 @@ defmodule Recode.CLIFormatterTest do
 
     output =
       capture_io(fn ->
-        format({:tasks_finished, project})
+        format({:tasks_finished, project, 10_000})
       end)
 
     assert output |> strip_esc_seq() |> String.trim_leading() == """
            File: foo.ex
            New file, created by Test
 
+           Executed 0 tasks in 0.01s.
            Files: 1 (.ex: 1)
            Created 1 file
             Everything ok
@@ -291,7 +302,7 @@ defmodule Recode.CLIFormatterTest do
 
     output =
       capture_io(fn ->
-        format({:tasks_finished, project})
+        format({:tasks_finished, project, 10_000})
       end)
 
     output = strip_esc_seq(output)
@@ -301,6 +312,7 @@ defmodule Recode.CLIFormatterTest do
            [foo 1/2] do not do this
            [bar 2/3] no no no
 
+           Executed 0 tasks in 0.01s.
            Files: 1 (.ex: 1)
             Found 2 issues
            """
@@ -326,7 +338,7 @@ defmodule Recode.CLIFormatterTest do
 
     output =
       capture_io(fn ->
-        format({:tasks_finished, project})
+        format({:tasks_finished, project, 10_000})
       end)
 
     output = strip_esc_seq(output)
@@ -337,6 +349,7 @@ defmodule Recode.CLIFormatterTest do
            [foo 2/3] no
            [foo 3/1] no
 
+           Executed 0 tasks in 0.01s.
            Files: 1 (.ex: 1)
             Found 3 issues
            """
@@ -362,7 +375,7 @@ defmodule Recode.CLIFormatterTest do
 
     output =
       capture_io(fn ->
-        format({:tasks_finished, project})
+        format({:tasks_finished, project, 10_000})
       end)
 
     output = strip_esc_seq(output)
@@ -373,6 +386,7 @@ defmodule Recode.CLIFormatterTest do
            [foo 2/2] no
            [foo 2/3] no
 
+           Executed 0 tasks in 0.01s.
            Files: 1 (.ex: 1)
             Found 3 issues
            """
@@ -398,7 +412,7 @@ defmodule Recode.CLIFormatterTest do
 
     output =
       capture_io(fn ->
-        format({:tasks_finished, project})
+        format({:tasks_finished, project, 10_000})
       end)
 
     output = strip_esc_seq(output)
@@ -415,6 +429,7 @@ defmodule Recode.CLIFormatterTest do
            Version 1/2 [foo 1/2] do not do this
            Version 1/2 [bar 2/3] no no no
 
+           Executed 0 tasks in 0.01s.
            Files: 1 (.ex: 1)
            Updated 1 file
             Found 2 issues
@@ -439,7 +454,7 @@ defmodule Recode.CLIFormatterTest do
 
     output =
       capture_io(fn ->
-        format({:tasks_finished, project})
+        format({:tasks_finished, project, 10_000})
       end)
 
     output = strip_esc_seq(output)
@@ -449,6 +464,7 @@ defmodule Recode.CLIFormatterTest do
            Execution of the Test task failed with error:
            Error Message
 
+           Executed 0 tasks in 0.01s.
            Files: 1 (.ex: 1)
             Found 1 issue
 
@@ -471,7 +487,7 @@ defmodule Recode.CLIFormatterTest do
         format({:prepared, project, 9999})
       end)
 
-    assert strip_esc_seq(output) == "Found 1 file\n"
+    assert strip_esc_seq(output) == "Read 1 file\n"
   end
 
   test "formats when tasks ready for an empty project" do
@@ -479,7 +495,7 @@ defmodule Recode.CLIFormatterTest do
 
     output =
       capture_io(fn ->
-        assert format({:tasks_finished, project})
+        assert format({:tasks_finished, project, 10_000})
       end)
 
     assert output == ""
@@ -498,10 +514,11 @@ defmodule Recode.CLIFormatterTest do
 
     output =
       capture_io(fn ->
-        assert format({:tasks_finished, project})
+        assert format({:tasks_finished, project, 10_000})
       end)
 
     assert output |> strip_esc_seq() |> String.trim_leading() == """
+           Executed 0 tasks in 0.01s.
            Files: 1 (.ex: 1)
             Everything ok
 
@@ -553,7 +570,7 @@ defmodule Recode.CLIFormatterTest do
 
     output =
       capture_io(fn ->
-        format({:task_finished, source, TaTask})
+        format({:task_finished, source, TaTask, 1})
       end)
 
     assert strip_esc_seq(output) == "."
@@ -573,7 +590,7 @@ defmodule Recode.CLIFormatterTest do
 
     output =
       capture_io(fn ->
-        format({:task_finished, source, TaTask})
+        format({:task_finished, source, TaTask, 1})
       end)
 
     assert strip_esc_seq(output) == "!"
@@ -593,7 +610,7 @@ defmodule Recode.CLIFormatterTest do
 
     output =
       capture_io(fn ->
-        format({:task_finished, source, TaTask})
+        format({:task_finished, source, TaTask, 1})
       end)
 
     assert strip_esc_seq(output) == "!"
@@ -610,10 +627,11 @@ defmodule Recode.CLIFormatterTest do
 
     output =
       capture_io(fn ->
-        format({:task_finished, source, TaTask}, debug: true)
+        format({:task_finished, source, TaTask, 10_000}, debug: true)
       end)
 
-    assert strip_esc_seq(output) == "Finished Elixir.TaTask with test/formatter_test.ex.\n"
+    assert strip_esc_seq(output) ==
+             "Finished Elixir.TaTask with test/formatter_test.ex [10000μs].\n"
   end
 
   defp format(message, opts \\ []) do